+++ WordPalette/WordPalette/Domain/Repository/UnsolvedWordRepository.swift
@@ -9,9 +9,18 @@ import Foundation
 import RxSwift
 
 protocol UnsolvedWordRepository {
-    func searchWord(keyword: String, level: Level) -> Single<[WordEntity]>
-    func saveWord(word: WordEntity) -> Single<Bool>
+    // JSON에서 전체 단어 로드
+    func fetchAllWords(level: Level) -> Single<[WordEntity]>
+    // DB에서 단어 로드
     func fetchWords(for level: Level) -> Single<[WordEntity]>
+    
+    // 검색 (JSON)
+    func searchWords(keyword: String, level: Level) -> Single<[WordEntity]>
+    
+    // 단어 저장
+    func saveWord(word: WordEntity) -> Single<Bool>
+
+    // 삭제
     func deleteWord(by id: UUID) -> Single<Bool>
     func deleteAllWords() -> Single<Bool>
 }
