+++ WordPalette/WordPalette/Domain/UseCase/AddWord/AddWordUseCase.swift
@@ -12,22 +12,16 @@ protocol AddWordUseCase {
     // 레벨별 전체 단어 로드 (json)
     func fetchAllWords(level: Level) -> Single<[WordEntity]>
     
-    // DB에 저장된 단어 로드
-    func fetchDBWords(level: Level) -> Single<[WordEntity]>
-    
     // 랜덤 추천 20개
     func recommendRandomWords(level: Level) -> Single<[WordEntity]>
     
-    // 검색 (json 기준)
-    func searchWords(keyword: String, level: Level) -> Single<[WordEntity]>
-    
+    // json+DB (refresh-저장 중복 확인용, 검색용)
+    func fetchAllWordsMerged(level: Level) -> Single<[WordEntity]>
+    func searchWordsMerged(keyword: String, level: Level) -> Single<[WordEntity]>
+
     // 단어 저장
     func saveWord(word: WordEntity) -> Single<Bool>
     
     // 중복 체크 (json + DB)
     func checkDuplicate(word: String) -> Single<(exists: Bool, level: Level?)>
-    
-    func fetchAllWordsMerged(level: Level) -> Single<[WordEntity]>
-    
-    func searchWordsMerged(keyword: String, level: Level) -> Single<[WordEntity]>
 }
